+++ cdk-examples/rust-bins/lib/rust-bins-stack.ts
@@ -24,6 +24,9 @@ export class RustBinsStack extends Stack {
         // binary, individually.
         // Settings.BUILD_INDIVIDUALLY = true;
 
+        // Uncomment to cross-compile Rust code to a different Lambda architecture.
+        // Settings.TARGET = 'aarch64-unknown-linux-gnu';
+
         const bucket = new s3.Bucket(this, 'RustBinsBucket', {
             blockPublicAccess: BlockPublicAccess.BLOCK_ALL,
             publicReadAccess: false,